refactor(ui): move TransitionLink to ui components and update button styles
Moved TransitionLink component to the ui components directory for better organization. Updated button styles to use classnames for dynamic class application and improved consistency in styling across the application. Also adjusted the theme colors to a black accent for a more modern look.
