+++ src/components/layout/Header.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import styles from './Header.module.scss';
-import { TransitionLink } from '../TransitionLink';
+import { TransitionLink } from '../ui/TransitionLink/TransitionLink';
 
 export const Header = () => {
   return (
