+++ src/components/ui/Button/Button.tsx
@@ -1,16 +1,17 @@
 import React from 'react';
 import styles from './Button.module.scss';
+import classNames from 'classnames';
 
 interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
   children: React.ReactNode;
+  className?: string;
 }
 
-const Button: React.FC<ButtonProps> = ({ children, ...props }) => {
+const Button: React.FC<ButtonProps> = ({ children, className, ...props }) => {
   return (
-    <button className={styles.button} {...props}>
+    <button className={classNames(styles.wrap, className)} {...props}>
       {children}
     </button>
   );
-};
-
+}
 export default Button;
