+++ src/components/ui/TransitionLink/TransitionLink.module.scss
@@ -1,4 +1,4 @@
-.wrap {
+.wrap{
   padding: 1rem 2rem;
   border-radius: var(--radius-md);
   background-color: var(--color-accent);
