+++ src/features/projects/components/ProjectHeader.module.scss
@@ -10,27 +10,5 @@
     display: flex;
     align-items: center;
     gap: 1rem;
-    & > * {
-      padding: 1rem 2rem;
-      border-radius: var(--radius-md);
-      box-shadow: var(--shadow-sm);
-      background-color: var(--color-accent);
-      color: var(--color-text-primary);
-      font-family: 'Pretendard', sans-serif;
-      font-size: var(--font-size-md);
-      line-height: 1.75;
-      transition:
-        background-color 0.2s ease,
-        color 0.2s ease,
-        transform 0.1s ease;
-      &:hover {
-        background-color: var(--color-accent-hover);
-      }
-
-      &:active {
-        transform: scale(0.96);
-        box-shadow: none;
-      }
-    }
   }
 }
