+++ src/features/projects/components/ProjectItem.module.scss
@@ -47,6 +47,7 @@
     flex-direction: column;
     gap: 0.75rem;
     flex: 1;
+    width: 100%;
   }
 
   .title {
