+++ src/styles/globals.scss
@@ -56,27 +56,11 @@ a {
 }
 
 button {
-  padding: 1rem 2rem;
-  border-radius: var(--radius-md);
-  box-shadow: var(--shadow-sm);
-  background-color: var(--color-accent);
-  color: var(--color-text-primary);
   font-family: 'Pretendard', sans-serif;
   font-size: var(--font-size-md);
   line-height: 1.75;
   cursor: pointer;
-  transition:
-    background-color 0.2s ease,
-    color 0.2s ease,
-    transform 0.1s ease;
-  &:hover {
-    background-color: var(--color-accent-hover);
-  }
 
-  &:active {
-    transform: scale(0.96);
-    box-shadow: none;
-  }
 }
 
 input,
